refactor: 프로젝트 코드 정리 및 최적화
## 삭제된 파일
- 중복 폴더 제거 (lib/data 2, lib/features 2, lib/shared 2)
- 사용하지 않는 auth_screen_old.dart 삭제
- 중복된 settings_screen.dart 삭제
- 루트의 빌드 생성 파일 삭제 (canvaskit, assets 등)

## 코드 수정
- AppColors에 warning 속성 추가
- AppHapticFeedback에 selection() 메서드 추가
- daily_reward_screen.dart 타입 에러 수정 (num -> int)
- practice_screen.dart null safety 문제 수정

## 프로젝트 설정
- .gitignore에 웹 빌드 아티팩트 추가
- assets 폴더 구조 재생성

+++ lib/features/daily_reward/daily_reward_screen.dart
@@ -342,7 +342,7 @@ class _DayRewardCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return TweenAnimationBuilder<double>(
-      duration: Duration(milliseconds: 300 + (reward.day * 50)),
+      duration: Duration(milliseconds: (300 + reward.day * 50).toInt()),
       curve: Curves.easeOutBack,
       tween: Tween(begin: 0.8, end: 1.0),
       builder: (context, scale, child) {
